requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.82.0
litellm>=1.30.0
langgraph>=0.0.10
dspy-ai>=2.6.24
pytest>=7.0.0
pytest-asyncio>=0.18.0
httpx>=0.24.0
python-multipart>=0.0.5
typing-extensions>=4.2.0
starlette>=0.27.0
requests>=2.31.0
aiohttp>=3.9.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0